server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Coders for Causes"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.2"
djangorestframework = "^3.14.0"
psycopg = {extras = ["binary", "pool"], version = "^3.1.9"}
djangorestframework-simplejwt = "^5.2.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
